fix: resolve implicit conversion warnings in lock_file and package_cache
- Fix integer precision loss in read_u16 by wrapping result cast
- Fix sign conversion in file.write by casting to std::streamsize

+++ src/dotvm/pkg/lock_file.cpp
@@ -52,7 +52,8 @@ void write_string(std::vector<std::uint8_t>& buf, std::string_view str) {
 
 /// @brief Read a 16-bit little-endian value
 [[nodiscard]] std::uint16_t read_u16(const std::uint8_t* data) {
-    return static_cast<std::uint16_t>(data[0]) | (static_cast<std::uint16_t>(data[1]) << 8);
+    return static_cast<std::uint16_t>(static_cast<std::uint16_t>(data[0]) |
+                                      static_cast<std::uint16_t>(data[1] << 8));
 }
 
 /// @brief Read a 32-bit little-endian value
